Remove 'cargo:rerun-if-changed=generator/'
The generator is a dependency crate, so Cargo will already handles it on its own
when in the Git repo.

This fixes the issue of the directory being marked as dirty when building
the unicode_names crate from crates.io rather than from the Git repo, as the
directory doesn't exist in .crate packages.